More updates.

+++ scripts/build.sh
@@ -1,6 +1,12 @@
 #!/usr/bin/env bash
 
+format=${1:-tex_pdf}
+
 cd dist || exit
-json_resume convert --out=tex_pdf --template=../templates/pdf.mustache ../rosshadden.json
-mv resume.pdf rosshadden.pdf
+if [ -f resume.pdf ]; then
+	json_resume convert --out="$format" --template=../templates/pdf.mustache ../rosshadden.json
+	mv resume.pdf rosshadden.pdf
+else
+	json_resume convert --out="$format" ../rosshadden.json
+fi
 cd .. || exit
